+++ src/helpers/comic_source.ts
@@ -1,38 +1,17 @@
-import type { Size } from "./size.ts";
-
 /**
  * Controls what contents are shown in the viewer. It can manage
  * error handling and throttling.
  *
  * @returns An array of whole page image sources.
  */
-export type ComicSource = (params: ComicSourceParams) => PromiseOrValue<MediaSourceOrDelay[]>;
+export type ComicSource = () => PromiseOrValue<MediaSourceResolver[]>;
+
+export type MediaSourceResolver = (params: SourceRefreshParams) => PromiseOrValue<MediaElement>;
 
 export type SourceRefreshParams = {
   /** The cause of the comic source being loaded. */
-  cause: "load" | "download" | "resize";
-  /** The page number to load. undefined for all pages. */
-  page?: number;
-} | {
-  /** The cause of the comic source being loaded. */
-  cause: "error";
-  /** The page number to load. undefined for all pages. */
-  page: number;
+  cause: "load" | "download" | "resize" | "error";
 };
-
-export type ComicSourceParams = SourceRefreshParams & {
-  /** Possible maximum viewer size until now. */
-  maxSize: Size;
-};
-
-/** `undefined` means delay the source loading. Viewer will request source again. */
-export type MediaSourceOrDelay = MediaSource | Delay;
-
-/** Provided remote image. */
-export type MediaSource = SimpleSource | MediaElement;
-type Delay = undefined;
-
-type SimpleSource = string;
 
 /**
  * Width and height are planned to be used for CLS prevention.
@@ -40,26 +19,12 @@ type SimpleSource = string;
  */
 export type MediaElement = HTMLImageElement | HTMLVideoElement;
 
-type PromiseOrValue<T> = T | Promise<T>;
+export type PromiseOrValue<T> = T | Promise<T>;
 
 export const MAX_RETRY_COUNT = 6;
 export const MAX_SAME_URL_RETRY_COUNT = 2;
 
-export function isDelay(sourceOrDelay: MediaSourceOrDelay): sourceOrDelay is Delay {
-  return sourceOrDelay === undefined;
-}
-
-export function toMediaElement(source: MediaSourceOrDelay): MediaElement {
-  if (isDelay(source)) {
-    return new Image();
-  }
-
-  if (typeof source === "string") {
-    const img = new Image();
-    img.src = source;
-    return img;
-  }
-
+export function normalizeMediaElement(source: MediaElement): MediaElement {
   if (source.isConnected) {
     return source.cloneNode(true) as MediaElement;
   }
@@ -68,33 +33,18 @@ export function toMediaElement(source: MediaSourceOrDelay): MediaElement {
 }
 
 export async function* getMediaIterable(
-  { media, index, comic, maxSize }: {
-    media: MediaSourceOrDelay;
-    index: number;
-    comic?: ComicSource;
-    maxSize: Size;
+  { media, initialCause = "load" }: {
+    media: MediaSourceResolver;
+    initialCause?: "load" | "download";
   },
 ) {
-  if (!isDelay(media)) {
-    yield getUrl(media);
-  }
-
-  if (!comic) {
-    return;
-  }
-
   let previous: string | undefined;
   let retryCount = 0;
   let sameUrlRetryCount = 0;
 
   while (sameUrlRetryCount <= MAX_SAME_URL_RETRY_COUNT && retryCount <= MAX_RETRY_COUNT) {
-    const hadError = media !== undefined || retryCount > 0;
-    const medias = await comic({ cause: hadError ? "error" : "load", page: index, maxSize });
-    const next = medias[index];
-    if (isDelay(next)) {
-      continue;
-    }
-
+    const cause = retryCount > 0 ? "error" : initialCause;
+    const next = await media({ cause });
     const url = getUrl(next);
     yield url;
 
@@ -107,6 +57,6 @@ export async function* getMediaIterable(
   }
 }
 
-function getUrl(source: MediaSource): string {
-  return typeof source === "string" ? source : source.src;
+function getUrl(source: MediaElement): string {
+  return normalizeMediaElement(source).src;
 }