+++ src/mod.tsx
@@ -3,7 +3,7 @@
 // @name:ko        vim comic viewer
 // @description    Universal comic reader
 // @description:ko 만화 뷰어 라이브러리
-// @version        20.3.0
+// @version        21.0.0
 // @namespace      https://greasyfork.org/en/users/713014-nanikit
 // @exclude        *
 // @match          http://unused-field.space/
@@ -46,7 +46,13 @@ import { createRoot, createStore, forwardRef, Provider, useMemo } from "./deps.t
 
 export type { ViewerController } from "./atoms/controller_atom.ts";
 export type { ViewerOptions } from "./atoms/viewer_base_atoms.ts";
-export type { ComicSource, ComicSourceParams, MediaSource } from "./helpers/comic_source.ts";
+export type {
+  ComicSource,
+  MediaElement,
+  MediaSourceResolver,
+  PromiseOrValue,
+  SourceRefreshParams,
+} from "./helpers/comic_source.ts";
 export { download } from "./helpers/downloader.ts";
 export * as utils from "./utils.ts";